[Refactor] 폴더구조 리빌딩
