+++ lib/information/view/index.dart
