+++ lib/journal/view/index.dart
