+++ lib/login/view/main_view_model.dart
@@ -1,4 +1,4 @@
-import 'package:fish_note/view/login/social_login.dart';
+import 'package:fish_note/login/view/social_login.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 
 class MainViewModel {
@@ -8,17 +8,16 @@ class MainViewModel {
 
   MainViewModel(this._socialLogin);
 
-  Future login() async{
+  Future login() async {
     isLogined = await _socialLogin.login();
-    if(isLogined){
+    if (isLogined) {
       user = await UserApi.instance.me();
-
     }
   }
-  Future logout() async{
+
+  Future logout() async {
     await _socialLogin.logout();
     isLogined = false;
     user = null;
-
   }
 }
