+++ lib/signUp/view/sign_up_affiliation.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:fish_note/components/next_button.dart';
+import 'package:fish_note/signUp/components/next_button.dart';
 import 'package:fish_note/theme/colors.dart';
 import 'package:fish_note/theme/font.dart';
 import 'package:flutter/material.dart';
